Remove a bunch of exposed methods from KV::Bucket
Don't know how useful they are yet. I added the getters because that's
just kinda what I've been doing with other resources, but buckets are
primarily intended to be used by applications and I don't know why you'd
need to know half this stuff for an application. I'm not even sure we
need to store it on the object. ¯\_(ツ)_/¯