Make subscription channels 64000 long by default
The Go client uses 65536 (2**16). I just kinda used 64000 to get
somewhere in the ballpark.

It turns out, 10 was too low a default for some scenarios. For example:

    nats.subscribe "request-me" do |msg|
      nats.reply msg, "response"
    end

    nats.subscribe "stuff" do |msg|
      nats.request "request-me"
    end

    50.times { nats.publish "stuff" }

What's happening is:

1. Receive 50 messages on the "stuff" subscription
2. First message, handler makes a request to "request-me"

The NATS::Client#begin_inbound method never makes it to the response in
step 2 because the subscription channel depth is only 10 but there are
50 messages. So `begin_inbound` is blocked on 40 messages that it can't
place into the subscription channel for "stuff" because it's full, but
there's only one fiber reading off the socket.

I don't *love* having to increase the channel size since it uses more
memory for every subscription (probably way more than needed for 98% of
them), but using more memory is better than broken functionality.