Naming changes for consistency and adds ability to publish a NATS::Message
There were interchangeable uses of "message", "body", and "data" referring
to the message payload data. This change conforms to the NATS protocol
which uses "message" as the encompassing message (headers + payload), and
"payload" as the primary data.

Also adds another `publish` method that takes a NATS::Message struct and
pipes it to the existing `publish` method