Fix a few issues with pull subscriptions
Previously, the implementation would create too many of what the NATS
server calls "waiting pulls", to the point where the server would stop
sending the responses. This commit fixes that and also returns more
quickly when a message is published during the poll, but not enough to
fill a complete batch.