Fix expected_last_subject_sequence argument
This was originally intended to be the way `KV::Client#update` was
implemented, but after I added the abstraction I forgot to use it. I'm
not entirely sure why it compiled up to now, though, because we were
never using the `Int` version of it, only ever the `Nil` version, so it
didn't even try to compile the overload of the method that took an `Int`
until this commit.