Remove unnecessary scoping
It's in the same namespace. Also, I'm not sure why I added the `API::V1`
namespace to begin with. I think it was a lesson I learned from previous
projects, but I don't think it was the right lesson here. A lot of the
API versioning in first-party NATS clients is based on server version,
which this client doesn't support yet.