Use `Message#data`
All of the first-party NATS clients use `data` as the property
representing the message's payload, so it makes sense to do that here,
too. Eventually, we'll deprecate the `body` method, but it's used
*everywhere*, so let's not do that just yet.

This commit also adds a `Message#data_string` method because most apps
deal with text or JSON inside messages.