Make Headers an alias of HTTP::Headers
This allows headers to be case-insensitive and allows passing multiple
instances of the same header multiple times. The first-party NATS Go
client specifically calls out that they base NATS headers on HTTP
headers, so this seems like the right way to go.

This may prevent apps that don't use the `NATS::Headers` abstraction
from compiling. I think that's an acceptable tradeoff. Coding against
the concrete type when there's an explicit alias is a handy shortcut
when prototyping, but it's not correct. Apps that have been using the
`NATS::Headers` alias the whole time, like our test suite does, will
continue to compile with no problem.

https://github.com/nats-io/nats.go/blob/ea3ef928233d820bcc1aa2d95a3055d4c79758a3/nats.go#L3834-L3836