Read message payloads as strings
Most messages that come in will be JSON-formatted strings. Creating a
string from a slice involves a copy operation, which requires an
additional heap allocation and makes the object larger. Converting a
string to a slice, on the other hand, is almost free — there are no heap
allocations or copying, just a single stack-allocated object that
contains a pointer to the string's buffer.

This PR makes the canonical storage a string to get that performance
boost for messages containing strings. When you need the data as raw
bytes, the API is the same, but it's returning a new `Slice` every time.
This way, the `Message` object's minimum size is reduced.

In benchmarks on my laptop, this improves message throughput for small
string messages by up to 15% (approx 700k->800k per second) and larger
messages (approximately 512KB) by about 4% (approx 12k->12.5k per
second). This large an increase in throughput seems worth the tradeoff
of not storing the data in memory in the "correct" type.