Validate subjects for publishing and subscribing
I can't believe I hadn't added this yet. I only noticed because I ran
into it while base64-encoding something for a subject which was adding a
'\n' to the end of the subject. This is invalid for a NATS subject and
actually causes the NATS server to disconnect us because the protocol
line we sent was invalid. So this commit raises the exception before it
causes a problem for the entire program.